Add interactive Docker prompts and registry management features
- Introduced new traits for handling interactive prompts related to Docker environments, repositories, and deployment domains, enhancing user experience during configuration.
- Implemented Docker registry management functionalities, including login checks and credential handling for Amazon ECR and Azure ACR, improving compatibility and security.
- Refactored existing methods to utilize the new traits, streamlining the codebase and promoting reusability.
- Enhanced error handling and user feedback during the Docker login process, ensuring a smoother experience for users.
